requirements.txt
matplotlib
networkx
scapy
pillow~=12.0.0
pip~=25.2
qrcode~=8.2
django-qrcode~=0.3
colorama~=0.4.6
